reports/requirements.txt
xhtml2pdf
seaborn
python-dotenv
psycopg2
pandas
boto3